extract material icons to material slots
