+++ packages/grid/x-data-grid/src/components/index.ts
@@ -3,7 +3,7 @@ export * from './cell';
 export * from './containers';
 export * from './columnHeaders';
 export * from './columnSelection';
-export * from './icons';
+export * from '../material/icons';
 export * from './menu';
 export * from './panel';
 export * from './toolbar';
