+++ packages/grid/x-data-grid/src/constants/defaultGridSlotsComponents.ts
@@ -7,18 +7,11 @@ import MUIButton from '@mui/material/Button';
 import MUIIconButton from '@mui/material/IconButton';
 import MUITooltip from '@mui/material/Tooltip';
 import MUIPopper from '@mui/material/Popper';
-import { GridIconSlotsComponent, GridSlotsComponent } from '../models';
+import { GridSlotsComponent } from '../models';
 import {
-  GridArrowDownwardIcon,
-  GridArrowUpwardIcon,
   GridCell,
   GridSkeletonCell,
-  GridCheckIcon,
-  GridCloseIcon,
-  GridColumnIcon,
   GridColumnsPanel,
-  GridFilterAltIcon,
-  GridFilterListIcon,
   GridFilterPanel,
   GridFooter,
   GridLoadingOverlay,
@@ -27,68 +20,14 @@ import {
   GridPanel,
   GridPreferencesPanel,
   GridRow,
-  GridSaveAltIcon,
-  GridSeparatorIcon,
-  GridTableRowsIcon,
-  GridTripleDotsVerticalIcon,
-  GridViewHeadlineIcon,
-  GridViewStreamIcon,
-  GridMoreVertIcon,
-  GridExpandMoreIcon,
-  GridKeyboardArrowRight,
-  GridAddIcon,
-  GridRemoveIcon,
-  GridDragIcon,
   GridColumnHeaderFilterIconButton,
-  GridSearchIcon,
-  GridVisibilityOffIcon,
-  GridViewColumnIcon,
-  GridClearIcon,
-  GridLoadIcon,
 } from '../components';
 import { GridColumnMenu } from '../components/menu/columnMenu/GridColumnMenu';
-import { GridColumnUnsortedIcon } from '../components/columnHeaders/GridColumnUnsortedIcon';
 import { GridNoResultsOverlay } from '../components/GridNoResultsOverlay';
-
-const DEFAULT_GRID_ICON_SLOTS_COMPONENTS: GridIconSlotsComponent = {
-  BooleanCellTrueIcon: GridCheckIcon,
-  BooleanCellFalseIcon: GridCloseIcon,
-  ColumnMenuIcon: GridTripleDotsVerticalIcon,
-  OpenFilterButtonIcon: GridFilterListIcon,
-  FilterPanelDeleteIcon: GridCloseIcon,
-  ColumnFilteredIcon: GridFilterAltIcon,
-  ColumnSelectorIcon: GridColumnIcon,
-  ColumnUnsortedIcon: GridColumnUnsortedIcon,
-  ColumnSortedAscendingIcon: GridArrowUpwardIcon,
-  ColumnSortedDescendingIcon: GridArrowDownwardIcon,
-  ColumnResizeIcon: GridSeparatorIcon,
-  DensityCompactIcon: GridViewHeadlineIcon,
-  DensityStandardIcon: GridTableRowsIcon,
-  DensityComfortableIcon: GridViewStreamIcon,
-  ExportIcon: GridSaveAltIcon,
-  MoreActionsIcon: GridMoreVertIcon,
-  TreeDataCollapseIcon: GridExpandMoreIcon,
-  TreeDataExpandIcon: GridKeyboardArrowRight,
-  GroupingCriteriaCollapseIcon: GridExpandMoreIcon,
-  GroupingCriteriaExpandIcon: GridKeyboardArrowRight,
-  DetailPanelExpandIcon: GridAddIcon,
-  DetailPanelCollapseIcon: GridRemoveIcon,
-  RowReorderIcon: GridDragIcon,
-  QuickFilterIcon: GridSearchIcon,
-  QuickFilterClearIcon: GridCloseIcon,
-  ColumnMenuHideIcon: GridVisibilityOffIcon,
-  ColumnMenuSortAscendingIcon: GridArrowUpwardIcon,
-  ColumnMenuSortDescendingIcon: GridArrowDownwardIcon,
-  ColumnMenuFilterIcon: GridFilterAltIcon,
-  ColumnMenuManageColumnsIcon: GridViewColumnIcon,
-  ColumnMenuClearIcon: GridClearIcon,
-  LoadIcon: GridLoadIcon,
-  AddFilterIcon: GridAddIcon,
-  ColumnReorderIcon: GridDragIcon,
-};
+import materialSlots from '../material';
 
 export const DATA_GRID_DEFAULT_SLOTS_COMPONENTS: GridSlotsComponent = {
-  ...DEFAULT_GRID_ICON_SLOTS_COMPONENTS,
+  ...materialSlots,
   BaseCheckbox: MUICheckbox,
   BaseTextField: MUITextField,
   BaseFormControl: MUIFormControl,
