+++ packages/grid/x-data-grid/src/material/index.ts
@@ -1,0 +1,71 @@
+import type { GridIconSlotsComponent } from '../models';
+import {
+  GridAddIcon,
+  GridArrowDownwardIcon,
+  GridArrowUpwardIcon,
+  GridCheckIcon,
+  GridCloseIcon,
+  GridColumnIcon,
+  GridDragIcon,
+  GridExpandMoreIcon,
+  GridFilterAltIcon,
+  GridFilterListIcon,
+  GridKeyboardArrowRight,
+  GridMoreVertIcon,
+  GridRemoveIcon,
+  GridSaveAltIcon,
+  GridSearchIcon,
+  GridSeparatorIcon,
+  GridTableRowsIcon,
+  GridTripleDotsVerticalIcon,
+  GridViewHeadlineIcon,
+  GridViewStreamIcon,
+  GridVisibilityOffIcon,
+  GridViewColumnIcon,
+  GridClearIcon,
+  GridLoadIcon,
+} from './icons';
+import { GridColumnUnsortedIcon } from './icons/GridColumnUnsortedIcon';
+
+const iconSlots: GridIconSlotsComponent = {
+  BooleanCellTrueIcon: GridCheckIcon,
+  BooleanCellFalseIcon: GridCloseIcon,
+  ColumnMenuIcon: GridTripleDotsVerticalIcon,
+  OpenFilterButtonIcon: GridFilterListIcon,
+  FilterPanelDeleteIcon: GridCloseIcon,
+  ColumnFilteredIcon: GridFilterAltIcon,
+  ColumnSelectorIcon: GridColumnIcon,
+  ColumnUnsortedIcon: GridColumnUnsortedIcon,
+  ColumnSortedAscendingIcon: GridArrowUpwardIcon,
+  ColumnSortedDescendingIcon: GridArrowDownwardIcon,
+  ColumnResizeIcon: GridSeparatorIcon,
+  DensityCompactIcon: GridViewHeadlineIcon,
+  DensityStandardIcon: GridTableRowsIcon,
+  DensityComfortableIcon: GridViewStreamIcon,
+  ExportIcon: GridSaveAltIcon,
+  MoreActionsIcon: GridMoreVertIcon,
+  TreeDataCollapseIcon: GridExpandMoreIcon,
+  TreeDataExpandIcon: GridKeyboardArrowRight,
+  GroupingCriteriaCollapseIcon: GridExpandMoreIcon,
+  GroupingCriteriaExpandIcon: GridKeyboardArrowRight,
+  DetailPanelExpandIcon: GridAddIcon,
+  DetailPanelCollapseIcon: GridRemoveIcon,
+  RowReorderIcon: GridDragIcon,
+  QuickFilterIcon: GridSearchIcon,
+  QuickFilterClearIcon: GridCloseIcon,
+  ColumnMenuHideIcon: GridVisibilityOffIcon,
+  ColumnMenuSortAscendingIcon: GridArrowUpwardIcon,
+  ColumnMenuSortDescendingIcon: GridArrowDownwardIcon,
+  ColumnMenuFilterIcon: GridFilterAltIcon,
+  ColumnMenuManageColumnsIcon: GridViewColumnIcon,
+  ColumnMenuClearIcon: GridClearIcon,
+  LoadIcon: GridLoadIcon,
+  AddFilterIcon: GridAddIcon,
+  ColumnReorderIcon: GridDragIcon,
+};
+
+const materialSlots = {
+  ...iconSlots,
+};
+
+export default materialSlots;
